Improve importation progress text
Much useful when it comes to big playlists, when updating infos every 10 tracks,
sometimes it can feel like something is going wrong because it takes too long.
This way we can precisely know the progress.
It is also more pleasent when it's a single line updating rather than several lines on the terminal

Signed-off-by: MLZ94 <zugutoroberto@gmail.com>